pyproject.toml
[project]
name = "lwsadmin"
version = "0.1.0"
description = "An administrative backend for managing a Monero Light Wallet service."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "monero>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
]
